Moves the startup logic to the app service.
